Bugfix: call MFSL_GetContext from commands_Cache_inode.c.
With Ganeshell, MFSL is initialized then cache_inode is. When used with
MFSL_ASYNC, this leads to a wrong pointer to mfsl_parameters because
cache_inode creates its own copy of it and passes it to MFSL.
We now call MFSL_GetContext from commands_Cache_inode in order to be
coherent between layers.

Plus, it appeared that when we only keep a pointer passed from cache_inode,
mfsl_param is seen uninitialized by ganeshell when used with
cache_inode. Now copy mfsl_param instead of just keeping pointer.